requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
google-genai>=1.29.0
google-cloud-storage>=2.10.0
google-auth>=2.23.4
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
python-multipart>=0.0.6

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

